moved interfaces into one file
