+++ client/src/pages/Course/index.tsx
@@ -2,8 +2,7 @@ import React, { useState, useEffect, useContext } from "react";
 import { useHistory } from "react-router-dom";
 import { RouteComponentProps } from "react-router";
 import { functions } from "config/firebase";
-import { CourseData } from "utils/courses";
-import { PostData } from "utils/posts";
+import { CourseData, PostData } from "utils/types";
 import { Button, Typography } from "@material-ui/core";
 import Navbar from "components/Navbar";
 import Post from "pages/Post";
