--- client/src/utils/comments.ts
@@ -1,5 +1,0 @@
-export interface CommentData {
-  author: string;
-  comment: string;
-  id?: string;
-}
