--- client/src/utils/courses.ts
@@ -1,7 +1,0 @@
-export interface CourseData {
-  id?: string;
-  code: string;
-  name: string;
-  description: string;
-  educator: string;
-}
