--- client/src/utils/posts.ts
@@ -1,9 +1,0 @@
-export interface PostData {
-  title: string;
-  author: string;
-  description: string;
-  links: string;
-  id?: string;
-}
-
-// TODO: Delete Post => Cloud Function to delete associated comments
